Symmetry in legacy access control calculation
