--- src/lib/Archive/isPublicArchive.ts
@@ -1,5 +1,0 @@
-import { XyoArchive } from '@xyo-network/sdk-xyo-client-js'
-
-export const isPublicArchive = (archive?: XyoArchive | null): boolean => {
-  return !archive ? false : !archive.accessControl
-}
